use component for armory_week app, address comments
