+++ src/desktop/apps/armory_week/client/initializer.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { BannerPopUp } from 'desktop/apps/armory_week/components/BannerPopUp'
+import { BannerPopUp } from 'desktop/components/fair_week_marketing/BannerPopUp'
 
 export const init = () => {
   const bootstrapData = window.__BOOTSTRAP__
