+++ src/desktop/apps/frieze_week/index.js
@@ -1,9 +1,7 @@
-import React from 'react'
 import { renderLayout as _renderLayout } from '@artsy/stitch'
-
 import adminOnly from '../../lib/admin_only'
 import JSONPage from '../../components/json_page/es6'
-import PageScaffold from 'desktop/components/fair_week_marketing/PageScaffold'
+import { FairWeekPageScaffold } from 'desktop/components/fair_week_marketing/PageScaffold'
 import merge from 'lodash.merge'
 import queryString from 'query-string'
 
@@ -20,7 +18,6 @@ export class EditableFriezeWeekPage extends JSONPage {
   }
 
   async show(req, res, next) {
-    console.log('got to show???')
     try {
       if (req.query['m-id'] !== MARKETING_MODAL_ID) {
         const queryStringAsString = queryString.stringify(
@@ -39,7 +36,7 @@ export class EditableFriezeWeekPage extends JSONPage {
         },
         blocks: {
           head: '../../components/fair_week_marketing/meta.jade',
-          body: PageScaffold,
+          body: FairWeekPageScaffold,
         },
         locals: {
           assetPackage: 'banner_pop_up',
