+++ src/desktop/components/fair_week_marketing/PageScaffold.js
@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types'
 import React from 'react'
 import styled, { ThemeProvider } from 'styled-components'
 
@@ -22,6 +23,8 @@ const SectionTitle = Title.extend`
 
 const IntroductionText = Text.extend`
   line-height: 31px;
+  margin-bottom: 20px;
+  color: ${colors.grayDark};
   @media (max-width: 24em) {
     font-size: 20px;
     line-height: 26px;
@@ -37,7 +40,9 @@ const FairLogo = styled.img`
 `
 
 const ResponsiveRow = styled(Row)`
-  @media (max-width: 48em) {
+  ${props =>
+    props.paddingBottom &&
+    `padding-bottom: ${props.paddingBottom}px;`} @media (max-width: 48em) {
     margin-left: -8px;
     margin-right: -8px;
   }
@@ -66,7 +71,7 @@ const theme = {
   },
 }
 
-export default ({
+export const FairWeekPageScaffold = ({
   introduction,
   fair_coverage,
   event,
@@ -75,39 +80,34 @@ export default ({
 }) => (
   <ThemeProvider theme={theme}>
     <Container>
-      <ResponsiveRow style={{ paddingBottom: 50 }}>
+      <ResponsiveRow paddingBottom={50}>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle
             titleSize="large"
             dangerouslySetInnerHTML={{ __html: introduction.title }}
           />
         </Col>
+
         <Col lg={8} md={8} sm={12} xs={12}>
           <ReveredColumnOnMobile>
-            <IntroductionText
-              textSize="xlarge"
-              color={colors.grayDark}
-              style={{ marginBottom: 20 }}
-            >
+            <IntroductionText textSize="xlarge">
               {introduction.description}
             </IntroductionText>
-            <div>
-              <img
-                style={{ marginTop: 30, marginBottom: 20, maxWidth: '100%' }}
-                src={introduction.image}
-              />
-            </div>
+            <img
+              style={{ marginTop: 30, marginBottom: 20, maxWidth: '100%' }}
+              src={introduction.image}
+            />
           </ReveredColumnOnMobile>
         </Col>
       </ResponsiveRow>
 
-      <ResponsiveRow style={{ paddingBottom: 50 }}>
+      <ResponsiveRow paddingBottom={50}>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle titleSize="large">{fair_coverage.title}</SectionTitle>
         </Col>
         <Col lg={8} md={8} sm={12} xs={12}>
-          <ResponsiveRow style={{ marginBottom: 20 }}>
-            {fair_coverage.fairs.map((fair) => (
+          <ResponsiveRow paddingBottom={20}>
+            {fair_coverage.fairs.map(fair => (
               <Col lg={3} md={3} sm={3} xs={6} key={fair.logo_url}>
                 {fair.site_url && fair.site_url.startsWith('http') ? (
                   <a href={fair.site_url} target="_blank">
@@ -122,7 +122,7 @@ export default ({
         </Col>
       </ResponsiveRow>
 
-      <ResponsiveRow style={{ paddingBottom: 45 }}>
+      <ResponsiveRow paddingBottom={45}>
         <Col lg={4} md={4} sm={12} xs={12}>
           <SectionTitle titleSize="large">{event.title}</SectionTitle>
         </Col>
@@ -154,8 +154,8 @@ export default ({
           </SectionTitle>
         </Col>
         <Col lg={8} md={8} sm={12} xs={12}>
-          {prepare_for_fairs.articles.map((article) => (
-            <ResponsiveRow style={{ marginBottom: 25 }} key={article.title}>
+          {prepare_for_fairs.articles.map(article => (
+            <ResponsiveRow paddingnBottom={25} key={article.title}>
               <Col lg={7} md={7} sm={6} xs={12}>
                 <a href={article.article_url} target="_blank">
                   <img
@@ -186,7 +186,15 @@ export default ({
         </Col>
       </ResponsiveRow>
 
-      {displayStickyFooter ? <div id="react-root-for-cta" /> : ''}
+      {displayStickyFooter && <div id="react-root-for-cta" />}
     </Container>
   </ThemeProvider>
 )
+
+FairWeekPageScaffold.propTypes = {
+  introduction: PropTypes.object,
+  fair_coverage: PropTypes.object,
+  event: PropTypes.object,
+  prepare_for_fairs: PropTypes.object,
+  displayStickyFooter: PropTypes.bool,
+}
